requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
ctkmessagebox==2.5
    # via -r requirements.in
customtkinter==5.2.1
    # via
    #   -r requirements.in
    #   ctkmessagebox
darkdetect==0.8.0
    # via customtkinter
mysql-connector-python==8.2.0
    # via -r requirements.in
packaging==23.2
    # via
    #   -r requirements.in
    #   build
pillow==10.1.0
    # via
    #   -r requirements.in
    #   ctkmessagebox
pip-tools==7.3.0
    # via -r requirements.in
protobuf==4.21.12
    # via mysql-connector-python
pyproject-hooks==1.0.0
    # via build
screeninfo==0.8.1
    # via -r requirements.in
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
wheel==0.41.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
